pyproject.toml
[project]
name = "retrieval-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-experimental>=0.3.4",
    "langchain>=0.3.19",
    "langchain-cohere==0.4.4",
    "cohere>=5.12.0,<5.13.0", 
    "langchain-openai>=0.3.7",
    "qdrant-client>=1.13.2",
    "ragas==0.2.10",
    "nltk==3.8.1",
    "rank-bm25>=0.2.2",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.6.0",
    "pymupdf>=1.24.0",  
    "rapidfuzz>=3.6.2",
    "fastapi==0.116.1",
    "starlette==0.47.2",
    "uvicorn==0.35.0",
]
